requirements.txt
Flask==1.0.2
flasgger==0.9.1
google-cloud-tasks==0.3.0
